+++ .gitignore
@@ -1,3 +1,4 @@
 backend/venv
 frontend/app/node_modules
 backend/pyenv
+.env
